Add experimental vscode-python setting
There’s an ongoing experiment in the vscode-python extension.
The experiment is named `pythonDeprecatePythonPath` (or
`pythonDeprecatePythonPathcf`, it’s not exactly clear). The purpose
of that experiment is that a Python interpreter is no longer stored in
`settings.json` if the user selects it manually.

Instead, a new `python.defaultInterpreterPath` setting is introduced,
which causes a Python interpreter to be selected automatically unless
the user has already selected one specifically for the workspace.

This commit adds that setting to make sure that whoever participates
in that experiment is still going to have a decent OOBE when checking
out this project.